pyproject.toml
[tool.poetry]
name = "qf635-market-microstructure"
version = "0.1.0"
description = ""
authors = ["Brandon, JunHao, Lynn, Yohanes"]
readme = "README.md"
packages = [{ include = "bot" }]

[tool.poetry.dependencies]
python = "^3.10, <3.13"
absl-py = "2.1.0"
aiohappyeyeballs = "2.6.1"
aiohttp = "3.11.18"
aiosignal = "1.3.2"
alpaca-trade-api = "3.2.0"
appnope = "0.1.4"
asttokens = "3.0.0"
astunparse = "1.6.3"
async-timeout = "5.0.1"
attrs = "25.3.0"
certifi = "2024.8.30"
charset-normalizer = "3.4.0"
comm = "0.2.2"
contourpy = "1.3.1"
cycler = "0.12.1"
dateparser = "1.2.1"
debugpy = "1.8.9"
decorator = "5.1.1"
deprecation = "2.1.0"
et-xmlfile = "2.0.0"
exceptiongroup = "1.2.2"
executing = "2.1.0"
flatbuffers = "24.3.25"
fonttools = "4.55.2"
frozenlist = "1.6.0"
gast = "0.6.0"
google-pasta = "0.2.0"
grpcio = "1.68.1"
h5py = "3.12.1"
idna = "3.10"
ipykernel = "6.29.5"
ipython = "8.30.0"
jedi = "0.19.2"
joblib = "1.4.2"
jupyter-client = "8.6.3"
jupyter-core = "5.7.2"
keras = "3.7.0"
kiwisolver = "1.4.7"
libclang = "18.1.1"
lightgbm = "4.5.0"
markdown = "3.7"
markdown-it-py = "3.0.0"
markupsafe = "3.0.2"
matplotlib = "3.9.3"
matplotlib-inline = "0.1.7"
mdurl = "0.1.2"
ml-dtypes = "0.4.1"
msgpack = "1.0.3"
multidict = "6.4.3"
namex = "0.0.8"
nest-asyncio = "1.6.0"
numpy = "2.0.2"
openpyxl = "3.1.5"
opt-einsum = "3.4.0"
optree = "0.13.1"
orjson = "3.10.18"
packaging = "24.2"
pandas = "2.2.3"
parso = "0.8.4"
pexpect = "4.9.0"
pillow = "11.0.0"
platformdirs = "4.3.6"
prompt-toolkit = "3.0.48"
propcache = "0.3.1"
protobuf = "5.29.1"
psutil = "6.1.0"
ptyprocess = "0.7.0"
pure-eval = "0.2.3"
pybind11 = "2.13.6"
pycryptodome = "3.22.0"
pygments = "2.18.0"
pyparsing = "3.2.0"
python-binance = "1.0.28"
python-dateutil = "2.9.0.post0"
python-dotenv = "1.1.0"
python-math = "0.0.1"
pytz = "2024.2"
pyyaml = "6.0.1"
pyzmq = "26.2.0"
regex = "2024.11.6"
requests = "2.32.3"
rich = "13.9.4"
scikit-learn = "1.5.2"
scipy = "1.14.1"
seaborn = "0.13.2"
six = "1.17.0"
stack-data = "0.6.3"
tensorboard = "2.18.0"
tensorboard-data-server = "0.7.2"
tensorflow = "2.18.0"
tensorflow-io-gcs-filesystem = "0.37.1"
termcolor = "2.5.0"
threadpoolctl = "3.5.0"
tornado = "6.4.2"
traitlets = "5.14.3"
typing-extensions = "4.12.2"
tzdata = "2024.2"
tzlocal = "5.3.1"
urllib3 = "1.26.20"
uv = "0.6.17"
wcwidth = "0.2.13"
websocket-client = "1.8.0"
websockets = "10.4"
werkzeug = "3.1.3"
wrapt = "1.17.0"
yarl = "1.20.0"
polars = "^1.29.0"
pyarrow = "^20.0.0"
redis = "^6.1.0"
pydantic-settings = "^2.9.1"
loguru = "^0.7.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
